Move to HttpExtensions instead
