+++ src/Api/AdminConsole/Controllers/ValidationErrorTypedResultsExtensions.cs
@@ -1,4 +1,5 @@
 ﻿using Bit.Core.AdminConsole.Utilities.v2.Validation;
+using Bit.HttpExtensions;
 
 namespace Microsoft.AspNetCore.Http.HttpResults;
