+++ src/Api/AdminConsole/Models/Response/Organizations/VerifiedOrganizationDomainSsoDetailsResponseModel.cs
@@ -1,7 +1,7 @@
 ﻿// FIXME: Update this file to be null safe and then delete the line below
 #nullable disable
 
-using Bit.Core.Models.Api;
+using Bit.HttpExtensions;
 
 namespace Bit.Api.AdminConsole.Models.Response.Organizations;
