+++ src/Api/Auth/Controllers/EmergencyAccessController.cs
@@ -8,10 +8,10 @@ using Bit.Api.Auth.Models.Response;
 using Bit.Api.Vault.Models.Response;
 using Bit.Core.Auth.UserFeatures.EmergencyAccess;
 using Bit.Core.Exceptions;
-using Bit.Core.Models.Api;
 using Bit.Core.Repositories;
 using Bit.Core.Services;
 using Bit.Core.Settings;
+using Bit.HttpExtensions;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
