+++ src/Api/SecretsManager/Controllers/SecretsManagerEventsController.cs
@@ -4,11 +4,11 @@
 using Bit.Api.Dirt.Models.Response;
 using Bit.Api.Utilities;
 using Bit.Core.Exceptions;
-using Bit.Core.Models.Api;
 using Bit.Core.Models.Data;
 using Bit.Core.Repositories;
 using Bit.Core.SecretsManager.AuthorizationRequirements;
 using Bit.Core.SecretsManager.Repositories;
+using Bit.HttpExtensions;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
