+++ src/Api/Vault/Controllers/FoldersController.cs
@@ -4,10 +4,10 @@
 using Bit.Api.Vault.Models.Request;
 using Bit.Api.Vault.Models.Response;
 using Bit.Core.Exceptions;
-using Bit.Core.Models.Api;
 using Bit.Core.Services;
 using Bit.Core.Vault.Repositories;
 using Bit.Core.Vault.Services;
+using Bit.HttpExtensions;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
