+++ test/Api.IntegrationTest/AdminConsole/Controllers/OrganizationUserControllerBulkAutoConfirmTests.cs
@@ -9,11 +9,11 @@ using Bit.Core.AdminConsole.Enums;
 using Bit.Core.AdminConsole.Repositories;
 using Bit.Core.Billing.Enums;
 using Bit.Core.Enums;
-using Bit.Core.Models.Api;
 using Bit.Core.Models.Data;
 using Bit.Core.Models.Data.Organizations;
 using Bit.Core.Repositories;
 using Bit.Core.Services;
+using Bit.HttpExtensions;
 using NSubstitute;
 using Xunit;
