+++ test/Api.IntegrationTest/AdminConsole/Controllers/OrganizationUserControllerBulkRevokeTests.cs
@@ -10,9 +10,9 @@ using Bit.Core.AdminConsole.Providers.Interfaces;
 using Bit.Core.AdminConsole.Repositories;
 using Bit.Core.Billing.Enums;
 using Bit.Core.Enums;
-using Bit.Core.Models.Api;
 using Bit.Core.Models.Data;
 using Bit.Core.Repositories;
+using Bit.HttpExtensions;
 using Xunit;
 
 namespace Bit.Api.IntegrationTest.AdminConsole.Controllers;
