+++ test/Api.IntegrationTest/NotificationCenter/Controllers/NotificationsControllerTests.cs
@@ -11,6 +11,7 @@ using Bit.Core.NotificationCenter.Entities;
 using Bit.Core.NotificationCenter.Enums;
 using Bit.Core.NotificationCenter.Repositories;
 using Bit.Core.Repositories;
+using Bit.HttpExtensions;
 using Xunit;
 
 namespace Bit.Api.IntegrationTest.NotificationCenter.Controllers;
