+++ test/Api.IntegrationTest/SecretsManager/Controllers/SecretVersionsControllerTests.cs
@@ -5,9 +5,9 @@ using Bit.Api.IntegrationTest.SecretsManager.Helpers;
 using Bit.Api.SecretsManager.Models.Request;
 using Bit.Api.SecretsManager.Models.Response;
 using Bit.Core.Enums;
-using Bit.Core.Models.Api;
 using Bit.Core.SecretsManager.Entities;
 using Bit.Core.SecretsManager.Repositories;
+using Bit.HttpExtensions;
 using Xunit;
 
 namespace Bit.Api.IntegrationTest.SecretsManager.Controllers;
