+++ test/Api.IntegrationTest/SecretsManager/Controllers/SecretsControllerTests.cs
@@ -6,9 +6,9 @@ using Bit.Api.SecretsManager.Models.Request;
 using Bit.Api.SecretsManager.Models.Response;
 using Bit.Core.Entities;
 using Bit.Core.Enums;
-using Bit.Core.Models.Api;
 using Bit.Core.SecretsManager.Entities;
 using Bit.Core.SecretsManager.Repositories;
+using Bit.HttpExtensions;
 using Bit.Test.Common.Helpers;
 using Xunit;
