+++ test/Api.Test/Auth/Controllers/DevicesControllerTests.cs
@@ -5,9 +5,9 @@ using Bit.Core.Auth.UserFeatures.DeviceTrust;
 using Bit.Core.Context;
 using Bit.Core.Entities;
 using Bit.Core.Enums;
-using Bit.Core.Models.Api;
 using Bit.Core.Repositories;
 using Bit.Core.Services;
+using Bit.HttpExtensions;
 using Microsoft.Extensions.Logging;
 using NSubstitute;
 using Xunit;
